+++ application/plotlydash/dashboard.py
@@ -7,8 +7,8 @@ import dash_core_components as dcc
 from .layout import html_layout
 import pymysql
 import plotly.graph_objects as go
-
-db = pymysql.connect("localhost", "root", "", "gamedata")
+import time
+db = pymysql.connect("localhost", "root", "password", "gamedata")
 
 def create_dashboard(server):
     dash_app = dash.Dash(server=server,
@@ -32,26 +32,28 @@ def create_dashboard(server):
              multi=True,
             clearable=True
         ),
-            #create_data_table(df)
+           html.Button('Submit', id='button', type="submit", n_clicks=0), #create_data_table(df)
     ],
         id='dash-container'
     )
 
     @dash_app.callback(
                     dash.dependencies.Output('dash-container', 'children'),
-                    [dash.dependencies.Input('histogram-graph', 'value')])
-
-    def update_output(value):
-                                    query = pd.read_sql_query("SELECT * FROM pkmn WHERE Name in {}".format(value).replace("[", "(").replace("]" , ")"), db)
-                                    fig = go.Figure(data=[
-                                    go.Bar(name='Sp Atk', y=query['SpAtk'], x=value),
-                                    go.Bar(name='Sp Def', y=query['SpDef'], x=value),
-                                    go.Bar(name='Speed', y=query['Speed'], x=value),
-                                    go.Bar(name='Attack', y=query['Attack'], x=value),
-                                    go.Bar(name='Defense', y=query['Defense'], x=value),
-                                    go.Bar(name='HP', y=query['HP'], x=value)
-                            ])
-                                    fig.update_layout(barmode='group')
-                                    return fig.show()
-            
+                    [dash.dependencies.Input('histogram-graph', 'value')],
+                    [dash.dependencies.State('button', 'n_clicks')])
+    
+    def update_output(n_clicks,value):
+                                         
+                                            query = pd.read_sql_query("SELECT * FROM pkmn WHERE Name in {}".format(value).replace("[", "(").replace("]" , ")"), db)
+                                            fig = go.Figure(data=[
+                                            go.Bar(name='Sp Atk', y=query['SpAtk'], x=value),
+                                            go.Bar(name='Sp Def', y=query['SpDef'], x=value),
+                                            go.Bar(name='Speed', y=query['Speed'], x=value),
+                                            go.Bar(name='Attack', y=query['Attack'], x=value),
+                                            go.Bar(name='Defense', y=query['Defense'], x=value),
+                                            go.Bar(name='HP', y=query['HP'], x=value)
+                                    ])
+                                            fig.update_layout(barmode='group')
+                                            return fig.show()
+                                        
     return dash_app.server
