+++ application/plotlydash/layout.py
@@ -2,7 +2,7 @@ html_layout = '''<!DOCTYPE html>
                     <html>
                         <head>
                             {%metas%}
-                            <title>PokeCompare</title>
+                            <title>GameData</title>
                             {%favicon%}
                             {%css%}
                         </head>
@@ -11,7 +11,7 @@ html_layout = '''<!DOCTYPE html>
                               <div class="nav-wrapper">
                                 <a href="/">
                                     <img src="/static/img/logo.png" class="logo" />
-                                    <h1>PokeStat</h1>
+                                    <h1>Gamedata 2</h1>
                                   </a>
                                 <nav>
                                 </nav>
