+++ application/templates/index.jinja2
@@ -22,10 +22,10 @@
 </div>
 <div class="card">
   <img src="{{ url_for('static', filename='img/logo.png') }}" class="logo" />
-  <h1 class="site-title">Explore LoL 2019 World Cup Stats</h1>
-  <p class="subtitle">Take a look to results , used champs and my thing to the LoL 2019 WC</p>
-  <a href="/lol/" class="dash-link">
-    <span>Embdedded Plotly Dash</span>
+  <h1 class="site-title">Explore Rocket League Grafana Visualisations</h1>
+  <p class="subtitle">Take a look to results </p>
+  <a href="/rocket/" class="dash-link">
+    <span>Rocket League</span>
     <i class="fas fa-arrow-right"></i>
   </a>
 </div>
