+++ application/templates/rocket.jinja2
@@ -1,0 +1,24 @@
+{% extends 'layout.jinja2' %}
+
+{% block content %}
+    <div class="container">
+        <h1>{{title}}</h1>
+        <p>{{description}}</p>
+        <iframe src="http://localhost:3000/d/PDR80ugMz/rocket-league-dashboard?orgId=1"
+  name="thumbnails"
+  frameborder="0"
+  style="
+    position: fixed;
+    top: 0px;
+    bottom: 0px;
+    right: 0px;
+    width: 100%;
+    border: none;
+    margin: 0;
+    padding: 0;
+    overflow: hidden;
+    z-index: 999999;
+    height: 100%;
+  ">
+    </div>
+{% endblock %}
